+++ src/trading/OrderBook.cpp
@@ -2,11 +2,11 @@
 #include <algorithm>
 #include <json.hpp>
 
-void OrderBook::addBid(double price, long volume) {
+void OrderBook::addBid(double price, double volume) {
     bids[price].emplace_back(price, volume);
 }
 
-void OrderBook::addAsk(double price, long volume) {
+void OrderBook::addAsk(double price, double volume) {
     asks[price].emplace_back(price, volume);
 }
 
@@ -31,7 +31,7 @@ bool OrderBook::matchOrders() {
         Order& bid = bidQueue.front();
         Order& ask = askQueue.front();
 
-        long tradeVolume = std::min(bid.volume, ask.volume);
+        double tradeVolume = std::min(bid.volume, ask.volume);
         bid.volume -= tradeVolume;
         ask.volume -= tradeVolume;
 
@@ -54,7 +54,7 @@ void OrderBook::setOrderBook(const nlohmann::json& json) {
         for (const auto& entry : json["bids"]) {
             if (entry.size() >= 2) {
                 double price = entry[0].get<double>();
-                long volume = entry[1].get<long>();
+                double volume = entry[1].get<double>();
                 bids[price].emplace_back(price, volume);
             }
         }
@@ -64,7 +64,7 @@ void OrderBook::setOrderBook(const nlohmann::json& json) {
         for (const auto& entry : json["asks"]) {
             if (entry.size() >= 2) {
                 double price = entry[0].get<double>();
-                long volume = entry[1].get<long>();
+                double volume = entry[1].get<double>();
                 asks[price].emplace_back(price, volume);
             }
         }
